Create blank.yml
The Python program I wrote generates 9 graphs. The program downloads the raw data directly from the ECDC website. This data gives the daily new cases and daily new death for every day since Dec-31 for every country. The first two graphs are just the new cases and new deaths, I have put them here. The next two graphs are total cases and total deaths for each country. These are interesting, but don't take population into account. The next two are cases per 1M pop. and deaths per 1M pop. These are very interesting. If you look at the cases / 1M you get four groups. Italy and Spain, very bad. Australia, very good. UK and Sweden, not too bad and then all the other countries, including Ireland, France, US and Germany, which aren't doing great.
But this is very misleading because the number of positive cases you get is proportional to the amount of testing you do. A much better thing to look at is the number of deaths / 1M pop. Now you see more of the real picture. Again, Italy and Spain is very bad, but France is not far behind them and the UK not far behind that. Sweden is next worse, then we have US & Ireland, then Portugal, then Germany, quite good and lastly Australia, very good. Germany comes out much better than it looks in cases because it's doing much more testing. The UK comes out much worse than  it looks in cases, because it's doing much less testing than Ireland or the US. Note that the UK, bizarrely, only counts hospital deaths in it's official death figures, it doesn't count nursing homes. So far, about 49% of Irish deaths have been outside of hospitals, so the true UK death rate could be twice as bad as shown here.
The next graph is related to those two, the Case Fatality Rate. The ratio of deaths to cases. If you catch Covid-19, the experts think that you chance of dying is somewhere between about 0.1% and about 1%, so a CFR of 15%, like you see in France, means that the true number infections is 15 times to 150 times higher than the number of positive cases, so for France, that is not 100,000 but somewhere between 1.5M and 15M!
The next two graphs are the rate of increase in Cases and the rate of increase in Deaths. These are the ones we need to get down to zero, especially the latter one.